# server/docs/openapi.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a server for a financial transactions management application.",
    "version": "1.0.0",
    "title": "Budget Kept API"
  },
  "host": "192.168.1.167:3001",
  "basePath": "/",
  "tags": [
    {
      "name": "Users",
      "description": "Operations related to users"
    },
    {
      "name": "Transactions",
      "description": "Operations related to transactions"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/register": {
      "post": {
        "tags": ["Users"],
        "summary": "Register a new user",
        "description": "Registers a new user and initializes financials.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User registration details",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["username", "password"],
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "Request body incomplete"
          },
          "409": {
            "description": "User already exists"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": ["Users"],
        "summary": "Login a user",
        "description": "Logs in a user and returns a JWT token.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User login details",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["username", "password"],
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "400": {
            "description": "Request body incomplete or password does not match"
          },
          "404": {
            "description": "User does not exist"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/forgot-password": {
      "post": {
        "tags": ["Users"],
        "summary": "Initiate password reset",
        "description": "Initiates a password reset by sending a reset pin.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Username for password reset",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["username"],
              "properties": {
                "username": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reset pin sent"
          },
          "400": {
            "description": "Request body incomplete"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/reset-password": {
      "post": {
        "tags": ["Users"],
        "summary": "Reset user password",
        "description": "Resets the user's password using a reset pin.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Reset password details",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["username", "pin", "newPassword"],
              "properties": {
                "username": {
                  "type": "string"
                },
                "pin": {
                  "type": "string"
                },
                "newPassword": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password reset successfully"
          },
          "400": {
            "description": "Invalid or expired PIN"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/transactions": {
      "get": {
        "tags": ["Transactions"],
        "summary": "Get user transactions",
        "description": "Gets all transactions for the authenticated user.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Transactions retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/add-transaction": {
      "post": {
        "tags": ["Transactions"],
        "summary": "Add a new transaction",
        "description": "Adds a new transaction for the authenticated user.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Transaction details",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["type", "amount"],
              "properties": {
                "type": {
                  "type": "string"
                },
                "amount": {
                  "type": "number"
                },
                "category": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "date": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Transaction added successfully"
          },
          "400": {
            "description": "Missing required transaction fields"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/update-transaction": {
      "put": {
        "tags": ["Transactions"],
        "summary": "Update an existing transaction",
        "description": "Updates an existing transaction for the authenticated user.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Updated transaction details",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["transactionId", "type", "amount"],
              "properties": {
                "transactionId": {
                  "type": "string"
                },
                "type": {
                  "type": "string"
                },
                "amount": {
                  "type": "number"
                },
                "category": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "date": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction updated successfully"
          },
          "400": {
            "description": "Missing required transaction fields"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/remove-transaction": {
      "delete": {
        "tags": ["Transactions"],
        "summary": "Remove an existing transaction",
        "description": "Removes an existing transaction for the authenticated user.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Transaction ID",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["transactionId"],
              "properties": {
                "transactionId": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction removed successfully"
          },
          "400": {
            "description": "Transaction ID required"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Transaction not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/balance": {
      "get": {
        "tags": ["Transactions"],
        "summary": "Get user balance",
        "description": "Gets the current balance for the authenticated user.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Balance retrieved successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/verify-token": {
      "get": {
        "tags": ["Users"],
        "summary": "Verify JWT token",
        "description": "Verifies the JWT token.",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Token verified successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  },
  "definitions": {}
}